services/batch/requirements.txt
celery
cassandra-driver
redis
requests